Data-Generation/Conversation-Generation/requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
openai>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
ijson>=3.2.0

# Testing dependencies (optional)
flask>=3.0.0
pytest>=7.4.0
